requirements.txt
pandas
faiss-cpu
google-generativeai
numpy
fastapi
uvicorn[standard]
python-dotenv
pydantic
PyMuPDF
requests
gunicorn 